+++ app/src/main/java/org/simple/clinic/datepicker/DatePickerKeyFactory.kt
@@ -5,6 +5,7 @@ import kotlinx.parcelize.Parcelize
 import org.simple.clinic.appconfig.Country
 import org.simple.clinic.datepicker.calendar.CalendarDatePicker
 import org.simple.clinic.navigation.v2.ScreenKey
+import org.simple.clinic.textInputdatepicker.TextInputDatePickerSheet
 import java.time.LocalDate
 import javax.inject.Inject
 
@@ -12,11 +13,25 @@ class DatePickerKeyFactory @Inject constructor(
     private val country: Country
 ) {
 
-  // TODO: Return appropriate date picker depending on country
   fun key(
       preselectedDate: LocalDate,
       allowedDateRange: ClosedRange<LocalDate>
   ): ScreenKey {
+    return when (country.isoCountryCode) {
+      Country.BANGLADESH, Country.INDIA -> showCalendarDatePicker(preselectedDate, allowedDateRange)
+      Country.ETHIOPIA -> showTextInputDatePicker(preselectedDate, allowedDateRange)
+      else -> showCalendarDatePicker(preselectedDate, allowedDateRange)
+    }
+  }
+
+  private fun showTextInputDatePicker(preselectedDate: LocalDate, allowedDateRange: ClosedRange<LocalDate>): ScreenKey {
+    return TextInputDatePickerSheet.Key(prefilledDate = preselectedDate,
+        minDate = allowedDateRange.start,
+        maxDate = allowedDateRange.endInclusive
+    )
+  }
+
+  private fun showCalendarDatePicker(preselectedDate: LocalDate, allowedDateRange: ClosedRange<LocalDate>): ScreenKey {
     return CalendarDatePicker.Key(preselectedDate = preselectedDate,
         minDate = allowedDateRange.start,
         maxDate = allowedDateRange.endInclusive)